fix(type-graphql): `@Subscription` `subscribe` invalid type definition
The type of `SubscribeOption.subscribe` is set to `ResolverFn`. But it turns out that, when running, the GraphQL response for this subscription is an error. It was expected an `AsyncIterable` as return type, not `AsyncIterator`, as described in `ResolverFn`.
To solve this for this moment, I've forked the `type-graphql` project, changed this definition, and now it's working perfectly.
I've also joined the gitter channel to check if this behavior is correct.

https://gitter.im/type-graphql/Lobby