feat(OMDbService): Try different APIs to find movie data
If the movie data was not found by its title, tries to find it through OMDb's "search" API, which is more generic, and may solve problems with titles ending with ", The" suffixes.
If found, then gets the movie data by its ID, which is certain to find. This is needed because "search" API does not brings all the data we need.